Refactor tapisNextPageParam utility function
